Chage Regervation url
